upload PA#1 Integer Linear Programming
Redundant via insertion is highly effective in improving chip yield and reliability. In this assignment, we study the problem of double-cut via insertion (DVI) in a post-routing stage, where a single via can have, at most, one redundant via inserted next to it, each inserted redundant via can serve only one via, vias cannot be overlapped, and the goal is to insert as many redundant vias as possible. The DVI problem can be formulated as a 0-1 integer linear program (ILP).